Spot/requirements.txt
amqp==2.1.4
billiard==3.5.0.2
boto==2.38.0
celery==4.0.2
Django==1.10.5
django-allauth==0.30.0
django-constance==2.0.0
django-model-utils==2.2
django-queryinspect==1.0.0
django-watchman==0.15.0
djangorestframework==3.5.3
dj-database-url==0.4.1
drf-nested-routers==0.90.0
enum-compat==0.0.2
enum34==1.1.6
eventlet==0.20.1
Fabric==1.14.0
greenlet==0.4.12
gunicorn==19.6.0
itsdangerous==0.24
joblib==0.11
jsonfield==1.0.3
kombu==4.0.2
mock==2.0.0
numpy==1.12.0
psycopg2==2.6.2
pycurl==7.43.0
python-http-client==2.2.1
python-dotenv==0.7.1
pytz==2016.10
redis==2.10.5
scipy==0.19.0
scikit-learn==0.18.1
sendgrid==3.6.3
sendgrid-django==4.0.4
six==1.10.0
vine==1.1.3
